コードフォーマットを変更した

+++ App/BhDefine/Functions/event/Common/CommonEvent.js
@@ -1,35 +1,35 @@
 (function() {
 
-  let NodeMVCBuilder = net.seapanda.bunnyhop.modelprocessor.NodeMVCBuilder;
-  let BhNodeID = net.seapanda.bunnyhop.model.node.attribute.BhNodeID;
+  let NodeMvcBuilder = net.seapanda.bunnyhop.modelprocessor.NodeMvcBuilder;
+  let BhNodeId = net.seapanda.bunnyhop.model.node.attribute.BhNodeId;
   let BhNodeState = net.seapanda.bunnyhop.model.node.BhNode.State;
   let ImitationBuilder = net.seapanda.bunnyhop.modelprocessor.ImitationBuilder;
-  let ImitationID = net.seapanda.bunnyhop.model.node.imitation.ImitationID;
+  let ImitationId = net.seapanda.bunnyhop.model.node.imitation.ImitationId;
   let bhCommon = {};
 
   // 入れ替わってWSに移ったノードを末尾に再接続する
   function appendRemovedNode(newNode, oldNode, manuallyRemoved, bhNodeHandler, bhUserOpeCmd) {
-
     let outerEnd = newNode.findOuterNode(-1);
-    if (outerEnd.canBeReplacedWith(oldNode) && !manuallyRemoved) {
+    if (outerEnd.canBeReplacedWith(oldNode)
+        && !manuallyRemoved
+        && !newNode.equals(outerEnd)) { // return などの外部ノードを持たないノードは newNode == outerEnd となる.
       bhNodeHandler.replaceChild(outerEnd, oldNode, bhUserOpeCmd);
       bhNodeHandler.deleteNode(outerEnd, bhUserOpeCmd);
-      }
+    }
   }
 
   /**
    * 引数で指定したノードを作成し, ワークスペースに追加する
-   * @param bhNodeID 作成するノードのID
+   * @param bhNodeId 作成するノードのID
    * @param pos 追加時のワークスペース上の位置
    * @param bhNodeHandler ノード操作用オブジェクト
    * @param bhNodeTemplates ノードテンプレート管理オブジェクト
    * @param bhUserOpeCmd undo/redo用コマンドオブジェクト
    * @return 新規作成したノード
    * */
-  function addNewNodeToWS(bhNodeID, workspace, pos, bhNodeHandler, bhNodeTemplates, bhUserOpeCmd) {
-
-    let newNode = bhNodeTemplates.genBhNode(BhNodeID.create(bhNodeID), bhUserOpeCmd);
-    NodeMVCBuilder.build(newNode);
+  function addNewNodeToWS(bhNodeId, workspace, pos, bhNodeHandler, bhNodeTemplates, bhUserOpeCmd) {
+    let newNode = bhNodeTemplates.genBhNode(BhNodeId.create(bhNodeId), bhUserOpeCmd);
+    NodeMvcBuilder.build(newNode);
     bhNodeHandler.addRootNode(workspace, newNode, pos.x, pos.y, bhUserOpeCmd);
     return newNode;
   }
@@ -43,7 +43,6 @@
    * @param bhUserOpeCmd undo/redo用コマンドオブジェクト
    * */
   function replaceDescendant(rootNode, descendantPath, newNode, bhNodeHandler, bhUserOpeCmd) {
-
     let oldNode = rootNode.findSymbolInDescendants(descendantPath);
     bhNodeHandler.replaceChild(oldNode, newNode, bhUserOpeCmd);
     bhNodeHandler.deleteNode(oldNode, bhUserOpeCmd);
@@ -59,7 +58,6 @@
    * @param bhUserOpeCmd undo/redo用コマンドオブジェクト
    * */
   function moveDescendant(from, to, descendantPath, bhNodeHandler, bhUserOpeCmd) {
-
     let childToBeMoved = from.findSymbolInDescendants(descendantPath);
     let oldNode = to.findSymbolInDescendants(descendantPath);
     bhNodeHandler.replaceChild(oldNode, childToBeMoved, bhUserOpeCmd);
@@ -74,7 +72,6 @@
    * @param userOpeCmd undo/redo用コマンドオブジェクト
    * */
   function replaceStatWithNewStat(oldStat, newStat, bhNodeHandler, bhUserOpeCmd) {
-
     let nextStatOfOldStat = oldStat.findSymbolInDescendants('*', 'NextStat', '*');
     let nextStatOfNewStat = newStat.findSymbolInDescendants('*', 'NextStat', '*');
     bhNodeHandler.exchangeNodes(nextStatOfOldStat, nextStatOfNewStat, bhUserOpeCmd);
@@ -88,7 +85,6 @@
    * @param userOpeCmd undo/redo用コマンドオブジェクト
    */
   function reconnectOuter(node, candidates, bhMsgService, bhNodeHandler, userOpeCmd) {
-
     // 移動させる外部ノードを探す
     let nodeToReconnect = node.findOuterNode(1);
     if (nodeToReconnect === null)
@@ -107,12 +103,12 @@
     
     // 接続先が無い場合は, ワークスペースへ
     if (nodeToReplace === null) {
-      let posOnWS = bhMsgService.getPosOnWS(nodeToReconnect);
-      bhNodeHandler.moveToWS(nodeToReconnect.getWorkspace(), nodeToReconnect, posOnWS.x, posOnWS.y, userOpeCmd);
+      let posOnWS = bhMsgService.getPosOnWs(nodeToReconnect);
+      bhNodeHandler.moveToWs(nodeToReconnect.getWorkspace(), nodeToReconnect, posOnWS.x, posOnWS.y, userOpeCmd);
     }
     else {
-      let posOnWS = bhMsgService.getPosOnWS(nodeToReconnect);
-      bhNodeHandler.moveToWS(nodeToReconnect.getWorkspace(), nodeToReconnect, posOnWS.x, posOnWS.y, userOpeCmd);
+      let posOnWS = bhMsgService.getPosOnWs(nodeToReconnect);
+      bhNodeHandler.moveToWs(nodeToReconnect.getWorkspace(), nodeToReconnect, posOnWS.x, posOnWS.y, userOpeCmd);
       bhNodeHandler.exchangeNodes(nodeToReconnect, nodeToReplace, userOpeCmd);
     }
     return;
@@ -122,7 +118,6 @@
    * 外部ノードの接続先ノードを探す
    */
   function findNodeToBeReplaced(nodeToReconnect, nodeToCheckReplaceability, candidates) {
-
     let parent = nodeToCheckReplaceability.findParentNode();
     if (parent === null)
       return null;
@@ -265,7 +260,6 @@
    * any-Type のノード名と型を決定するシンボル名から対応する static-type Node のIDを取得する
    */
   function getStaticTypeNodeID(anyTypeNodeName, typeDecisiveSymbolName) {
-
     let sectionNameToStaticType = anyTypeToStaticTypeNode[anyTypeNodeName];
     if (!sectionNameToStaticType)
       return null;
@@ -281,7 +275,6 @@
    * static-type の式である場合 true を返す.
    */
   function isStaticTypeExp(node) {
-  
     let section = node.findSymbolInDescendants('*');
     let sectionName = null;
     if (section !== null)
@@ -302,16 +295,17 @@
    * @return node のイミテーションノード
    */
   function buildImitation(node, imitID, userOpeCmd) {
-    return ImitationBuilder.build(node, ImitationID.create(imitID), false, userOpeCmd);
+    return ImitationBuilder.build(node, ImitationId.create(imitID), false, userOpeCmd);
   }
   
-  /** BhNode を新規作成する
+  /**
+   * BhNode を新規作成する
    * @param nodeID 作成するノードのID
    * @param bhNodeTemplates ノードテンプレート管理オブジェクト
    * @param bhUserOpeCmd ndo/redo用コマンドオブジェクト
    */
-  function genBhNode(bhNodeID, bhNodeTemplates, bhUserOpeCmd) {
-    return bhNodeTemplates.genBhNode(BhNodeID.create(bhNodeID), bhUserOpeCmd);
+  function genBhNode(bhNodeId, bhNodeTemplates, bhUserOpeCmd) {
+    return bhNodeTemplates.genBhNode(BhNodeId.create(bhNodeId), bhUserOpeCmd);
   }
 
   bhCommon['appendRemovedNode'] = appendRemovedNode;
